+++ src/screens/datasets/actions.ts
@@ -12,7 +12,7 @@ import { IdMap, MintPreferences } from "app/reducers";
 import { DateRange } from "screens/modeling/reducers";
 import { Region } from "screens/regions/reducers";
 import { MINT_PREFERENCES } from "config";
-import { DataCatalogAdapter } from "util/data-catalog-adapter";
+import { DataCatalogAdapter } from "util/datacatalog/data-catalog-adapter";
 
 export const DATASETS_VARIABLES_QUERY = "DATASETS_VARIABLES_QUERY";
 export const DATASETS_GENERAL_QUERY = "DATASETS_GENERAL_QUERY";
